feat(subscription-manager): update design to support secure ssr
With the previous, stateful manager there are cases that could cause SSR to be insecure. If two
requests came in at the same time, the selector could've ended up selecting data from the wrong
store.

BREAKING CHANGE: The api is now reworked to support each subscription being it's own usage of the
library. There is no longer a stateful component in the module due to the security implications.